// web/src/shared/api/generated/model/examBlueprintResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ExamDifficultyMixResult } from './examDifficultyMixResult';
import type { ExamTypeCountResult } from './examTypeCountResult';

export interface ExamBlueprintResult {
  id: string;
  subjectId: string;
  /** @nullable */
  unitId: string | null;
  typeCounts: ExamTypeCountResult[];
  difficultyMix: null | ExamDifficultyMixResult;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  questionCount: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  timeLimitMinutes: number | string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  passMark: number | string;
}